fixed warnings and filters

+++ Source/RewardManager.cpp
@@ -44,7 +44,7 @@ void RewardManager::CanAddBonusBall()
 	std::mt19937 gen(rd());
 	std::uniform_real_distribution<> distribBallChance(1, m_chanceOfGainABall);
 
-	int random_number = distribBallChance(gen);
+	int random_number = (int)distribBallChance(gen);
 
 	if (random_number == 1) {
 		AddBonusBall();
